+++ pandapower/test/api/test_file_io.py
@@ -398,5 +398,19 @@ def test_empty_geo_dataframe():
     assert_net_equal(net, net1)
 
 
+
+def test_json_io_with_characteristics(net_in):
+    c1 = pp.control.Characteristic.from_points(net_in, [(0, 0), (1, 1)])
+    c2 = pp.control.SplineCharacteristic.from_points(net_in, [(2, 2), (3, 4), (4, 5)])
+
+    net_out = pp.from_json_string(pp.to_json(net_in))
+    assert_net_equal(net_in, net_out)
+    assert "characteristic" in net_out.keys()
+    assert isinstance(net_out.characteristic.object.at[c1.index], pp.control.Characteristic)
+    assert isinstance(net_out.characteristic.object.at[c2.index], pp.control.SplineCharacteristic)
+    assert np.isclose(net_out.characteristic.object.at[c1.index](0.5), c1(0.5), rtol=0, atol=1e-12)
+    assert np.isclose(net_out.characteristic.object.at[c2.index](2.5), c2(2.5), rtol=0, atol=1e-12)
+
+
 if __name__ == "__main__":
     pytest.main([__file__, "-s"])